// frontend/src/api/model/categoryItemDto.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * YaoYao Dinner API
 * Restaurant management API
 * OpenAPI spec version: 1.0
 */

export interface CategoryItemDto {
  id: string;
  key: string;
  /** @nullable */
  name: string | null;
  /** @nullable */
  description: string | null;
  isAvailable: boolean;
}